+++ compat/filegen.c
@@ -79,7 +79,7 @@ int main(int argc, char *argv[]) {
 
     /* Write data_out to argv[3] */
     f = fopen(argv[3], "wb+");
-    if (fwrite(data_out, 1, (size_t) csize, f) == csize) {
+    if (fwrite(data_out, 1, (size_t) csize, f) == (uint32_t) csize) {
       printf("Wrote %s\n", argv[3]);
     } else {
       printf("Write failed");
@@ -90,7 +90,7 @@ int main(int argc, char *argv[]) {
     fseek(f, 0, SEEK_END);
     fsize = ftell(f);
     fseek(f, 0, SEEK_SET);
-    if (fread(data_out, 1, (size_t) fsize, f) == fsize) {
+    if (fread(data_out, 1, (size_t) fsize, f) == (uint32_t) fsize) {
       printf("Checking %s\n", argv[2]);
     } else {
       printf("Read failed");
