+++ tests/test_frame.c
@@ -109,7 +109,7 @@ static char* test_frame() {
 
   if (usermeta) {
     int content_len_ = blosc2_get_usermeta(schunk, &content_);
-    mu_assert("ERROR: bad usermeta length in frame", content_len_ == content_len);
+    mu_assert("ERROR: bad usermeta length in frame", (size_t) content_len_ == content_len);
     mu_assert("ERROR: bad usermeta data in frame", strncmp((char*)content_, content, content_len) == 0);
     free(content_);
     blosc2_update_usermeta(schunk, (uint8_t *) content2, (int32_t) content_len2, BLOSC2_CPARAMS_DEFAULTS);
@@ -145,7 +145,7 @@ static char* test_frame() {
 
   if (usermeta) {
     int content_len_ = blosc2_get_usermeta(schunk, &content_);
-    mu_assert("ERROR: bad usermeta length in frame", content_len_ == content_len2);
+    mu_assert("ERROR: bad usermeta length in frame", (size_t) content_len_ == content_len2);
     mu_assert("ERROR: bad usermeta data in frame", strncmp((char*)content_, content2, content_len2) == 0);
     free(content_);
     blosc2_update_usermeta(schunk, (uint8_t *) content3, (int32_t) content_len3, BLOSC2_CPARAMS_DEFAULTS);
@@ -202,7 +202,7 @@ static char* test_frame() {
 
   if (usermeta) {
     int content_len_ = blosc2_get_usermeta(schunk, &content_);
-    mu_assert("ERROR: bad usermeta length in frame", content_len_ == content_len3);
+    mu_assert("ERROR: bad usermeta length in frame", (size_t) content_len_ == content_len3);
     mu_assert("ERROR: bad usermeta data in frame", strncmp((char*)content_, content3, content_len3) == 0);
     free(content_);
   }
